% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{plot.heatmap}
\alias{plot.heatmap}
\title{make quick heatmap plot for a seurat object}
\usage{
plot.heatmap(seuratObject, anno.col = NULL, gene.mark = NULL,
  marker = NULL, duplicate = FALSE, top = 20)
}
\arguments{
\item{seuratObject}{a Seurat object, need to heve done the cluster step.}

\item{anno.col}{data.frame, annotation for clusters, rows are clusters same as seuratObject, columns are classification variables}

\item{gene.mark}{a series of genes, which will significantly marked in the final heatmap.}

\item{marker}{a series of genes, shows the row value of heatmap.}

\item{duplicate}{logic value. can the marker in all groups be duplicated, accept TRUE/FALSE. if FALSE, the program will maintain the only marker with maixmum avg_logFC.}

\item{top}{filter paragram for default marker generate.}
}
\value{
a heatmap of interesting markers in all clusters
}
\description{
make quick heatmap plot for a seurat object
}
\examples{

}
