% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singler_annotation.R
\name{seu.singler}
\alias{seu.singler}
\title{combine seurat with singler , make auto annotation for seurat cluster}
\usage{
seu.singler(seuratObject, ref = NULL, ref.count = TRUE,
  method = "cluster")
}
\arguments{
\item{seuratObject}{a seurat object}

\item{ref}{matrix of gene exression or the output of ref.make(), or you can add one of database built in singleR. default is DatabaseImmuneCellExpressionData().}

\item{ref.count}{logic value, if the expression values in ref are count data, FALSE means values in ref has been lognormalized.}

\item{method}{"cluster" or "single", used for singler annotation strategy.}
}
\value{
dataframe of seurat cluster annotaion
}
\description{
combine seurat with singler , make auto annotation for seurat cluster
}
\examples{
library(scRNAwrapper)
sample<-readRDS("~/work/2020.1.7/data/srt.final.RDS")
sample<-subset(sample,ident=0:3)
anno<-seu.singler(sample)
sample$singler<-Idents(sample)
levels(sample$singler)<-anno[,2]
DimPlot(sample,group.by = "singler",label = TRUE)

}
